remove double-fetch when using leading polling
When I changed the polling mechanism to run a fetch initially, I found there were two post-list fetches happening at once. It turns out the post-list-fetcher had another mechanism to fetch the initial query, which is to run "fetchNextPage" when we're on page zero. By removing this we should have the initial fetch handled by the polling's leading fetch request.