Signup: Remove unnecessary conditional in getFlowName
The conditional read: if the requested flow name is not valid and the
currentFlowName is equal to the defaultFlowName, return the defaultFlowName.

But immediately afterward it returns the currentFlowName anyway, which would be
the same thing if the requested flow name was valid or invalid or if the
currentFlowName was the defaultFlowName or not.